Rust code now holds a unique ptr to the GUI class
This is a huge step forward: we can now create a proper GUI class
and call functions directly on that to instruct the GUI to do anything,
while the GUI itself has a pointer to the Rust's EventLoopProxy to send
user events and thus control the engine. We finally have some nice,
clean and simple 2 way of communication.