Changed code so that the graph is responsible for zoom
This is still far from perfect, because this adds some "display-only" data
to the graph (i.e: zoom levels). This data will interact weirdly with
undo/redo, and we need to manually fix this. Ideally, visualization
of the graph should be the responsibility of a separate class, so that
editing (including undo/redo) and display are clearly separated.